trig_utils: fix tstat output on 32-bit platforms
struct ubx_tstat.cnt is unsigned long but was printed via the
PRIu64 based FILE_FMT/LOG_FMT. On 32-bit platforms (e.g. armv7)
this shifts the subsequent varargs, garbling the min/max/avg
values in both the csv profile and the log output. Cast cnt to
uint64_t at the two print sites.

+++ libubx/trig_utils.c
@@ -73,7 +73,7 @@ int tstat_fwrite(FILE *fp, struct ubx_tstat *stats)
 		ubx_ts_div(&stats->total, stats->cnt, &avg);
 
 		fprintf(fp, FILE_FMT,
-			stats->id, stats->cnt,
+			stats->id, (uint64_t)stats->cnt,
 			ubx_ts_to_us(&stats->min),
 			ubx_ts_to_us(&stats->max),
 			ubx_ts_to_us(&avg));
@@ -97,7 +97,7 @@ void tstat_log(const ubx_block_t *b, const struct ubx_tstat *stats)
 	ubx_ts_div(&stats->total, stats->cnt, &avg);
 
 	ubx_info(b, LOG_FMT,
-		 stats->id, stats->cnt,
+		 stats->id, (uint64_t)stats->cnt,
 		 ubx_ts_to_us(&stats->min),
 		 ubx_ts_to_us(&stats->max),
 		 ubx_ts_to_us(&avg));
